[nxp-lpc] Fixed long start-up issue.

+++ ports/nxp/supervisor/port_lpc17xx.h
@@ -36,6 +36,6 @@ extern void RIT_ClearInt(void);
 extern void RIT_Disable(void);
 extern void RIT_Disable(void);
 extern void RIT_Enable(void);
-extern uint64_t RIT_GetCounter(void);
+extern uint32_t RIT_GetCounter(void);
 
 #endif // PORTS_NXP_SUPERVISOR_PORT_LPC17XX_H_
